requirements.txt
joblib==1.1.0
matplotlib==3.5.1
numpy==1.22.3
pandas==1.4.2
Pillow==9.1.0
pyarrow==7.0.0
python_dateutil==2.8.2
scikit_learn==1.0.2
seaborn==0.11.2
streamlit==1.8.1
